MicrosoftAppIn/requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
pyodbc
msal
pydantic
python-multipart
aiofiles
email-validator
requests
passlib
python-jose
cryptography
pytest
httpx